main_hms/pyproject.toml
[tool.poetry]
name = "main-hms"
version = "0.1.0"
description = ""
authors = ["Pavel <samoesamokhin@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "4.2.9"
gunicorn = "23.0.0"
asgiref = "3.8.1"
packaging = "24.1"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
tzdata = "2024.2"
celery = "^5.4.0"
redis = "^5.1.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.7.0"
flower = "^2.0.1"
channels = "^4.1.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
django-elasticsearch-dsl = "^8.0"
pillow = "^11.0.0"
motor = "^3.6.0"
stripe = "^11.2.0"
hiredis = "^3.0.0"
django-redis = "^5.4.0"
django-debug-toolbar = "^4.4.6"
flake8 = "^7.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ["py39"]